Switch over to django.contrib.auth.views.login
Once upon a time, django.contrib.auth didn't have a built-in login view. As of
about half a decade ago, it does. This means we can kill the one in djblets,
which isn't nearly as well maintained.

This change switches us over to the new view and form, which works pretty
similarly to the old one. The major differences are in the naming of the error
fields, 'next' vs. 'next_page' (which I'll deal with once I stop using
djblets.auth.util.login_required), and CSRF protection.

Testing done:
* Tried to login with an invalid username and password and saw the right error.
* Tried to login after fiddling with the CSRF token and saw a 403.
* Logged in with a valid username and password and was redirected to
  /dashboard/
* Checked that when I accessed a page when logged out I was redirected to log
  in, and then once I logged in, I was redirected back to that page.

Reviewed at https://reviews.reviewboard.org/r/4931/