Switch to django's built-in login_required decorator.
There's no longer any reason for us to have our own login_required decorator,
and it's just more code for us to maintain. The upstream django one provides
everything we need (plus some nice checking that we didn't have).

This is a little bit ugly because of the name of the 'next_page' field on the
login page, but that will be a trivial cleanup once this goes in.

Testing done:
Checked that all the views which explicitly used login_required and a few of
the views that used check_login_required correctly redirected to the login page
with next_page set properly. Once there, checked that logging in redirected to
the correct page.

Reviewed at https://reviews.reviewboard.org/r/4932/