Don't show the asterisk next to "Reviewers" when closed.
All the "required field" asterisks on the page were created dynamically
by the inlineEditor except for the one next to "Reviewers." That one was
always there if the user had edit access, even if the review request was
closed and couldn't be edited anymore. This fixes the logic to hide it
in this case.