Scroll back to the top of the diff viewer when switching pages.
The diff viewer page's scroll position would stay the same (or as close to the
same as possible with all the changes) whenever switching revisions or pages.
Most of the time this is ok (if using the controls at the top of the page, one
doesn't really expect it to scroll), but using the page selector at the bottom
resulted in having to manually scroll back up to the top. This change will
scroll back up to the index_header anchor when using the bottom page selector.

Testing done:
Switched back and forth between two pages of a diff using both the top and
bottom selectors. Saw that in the latter case, I was scrolled back up to the
file index.

Reviewed at https://reviews.reviewboard.org/r/4979/