Update a bunch of stuff in the docs.
I just did a read-through of the docs to check for things that are out of date.
Aside from having to completely rewrite the 'post-review' page, and many of our
screenshots being old and out of date, there were a few text changes that
needed to be made:

* The reference to "Getting Started" in the codebase "Custom Forks" page wasn't
  correct.
* The "Unit Tests Fixtures" page still listed test_reviewrequests and didn't
  have any information about the new `TestCase` class.
* The minimum browser versions still said we supported IE7.
* The "What is Code Review" page had out of date information about the state of
  our support for post-commit review.
* There was a typo in the paragraph about the "Bugs" field.
* There was no documentation for the "Depends On" field.

Testing done:
Built user manual and codebase html and looked at the results.

Reviewed at https://reviews.reviewboard.org/r/4981/