Instantiate CodeMirror widgets on demand only.
This lightens MarkdownEditor to only instantiate a CodeMirror wiget as
needed. MarkdownEditor will now create it when show() is called, or when
the inlineEditor begins editing. It will then go away when hidden or
when editing is finished. It should reduce the memory footprint a little
bit. Though it's a bit hard to tell what the savings is, it looks like
it may be as much as 8-10MB.