Improve the location of the "base-after-navbar" template hook.
This template hook was in a weird position before. It's, in theory, a
useful hook when you want to place something immediately after the
navbar, flush with the navbar, but instead it was in a container further
down that had padding around it.

This change moves it to be right after the navbar, and adds a new block
that the admin UI needed to place its sidebar.