Fix regression in comment dialog dragging.
The way that the comment dialog was made draggable regressed, possibly due to
upgrading jQuery UI. The result was that the entire dialog except for the
resize grips was treated as draggable, making it impossible to use the mouse to
select text in the editor. Switching it to pass in a selector instead of a
jQuery makes it work properly again.

Testing done:
Opened a comment dialog on a diff and checked that I got the correct cursor on
the title of the dialog, and that I could only drag that dialog around by
clicking on the title. Checked that I could now select text in the editor.

Reviewed at https://reviews.reviewboard.org/r/5016/