Un-escape markdown in text emails.
There's not a lot of processing we care to do for markdown in text emails, but
the escape characters for auto-escaped descriptions and comments is annoying.
This adds some processing that de-escapes markdown text for the text email
templates.

Testing Done:
Looked at the review request, review, and reply preview-email pages.

Reviewed at https://reviews.reviewboard.org/r/5011/