Enable usage of django.contrib.messages.
The Django messages framework has improved quite a bit since the early
days, and is now a useful thing to have in the admin UI. It already
stored cookies with these messages, but they just built up and were
never used or cleared. By activating the context processor, the admin UI
can provide much nicer responses to actions, such as deleting objects.

We can probably make use of this in some good ways going forward as
well.

While activating this, I noticed that output in the admin UI was wrong.
Our HTML was quite broken, since we never tested it before. This has
been fixed.