Fix so users with apostrophes won't cause 500 errors.
It's possible to get a user with an apostrophe due to active directory. This
change allows the regex to recognize these users.

Testing Done:
I created a user with an apostrophe in the name and went through the steps to
reproduce the problem (going to users page, as well as trying to assign the
apostrophe user to a review). 500 errors were no longer thrown after my
changes.

Bugs closed: 2890

Reviewed at https://reviews.reviewboard.org/r/4606/