Reduce the size of the repo_revision field for FileAttachments.
The repo_revisision field for FileAttachments was too long for an
index to be formed on MySQL. This caused issues when upgrading to 2.0
beta 1. We don't really need a revision field this long (and only had it
this long to match the one in FileDiff, which also doesn't need to be
that long).

We now limit it to 64, which is the length of a sha256sum. It should be
long enough.