Factor out all common request fields for comment API resources.
There's a number of fields that are common between all the comment
resources' create/update requests. In order to reduce code duplication
and to make it easier to add or change these lists, I've pulled out
these common fields into a set of variables in BaseCommentResource that
can be reused.