Switch the Mercurial support to use the command line tool.
Our Mercurial support interfaced with the 'hg' module to cat files in a
repository, but this can easily be accomplished by the command line
tool. That tool is the correct API to use, not the module. More
importantly, it resolves a licensing issue due to the 'hg' module being
GPL, which was a terrible oversight that we want to address. This will
be backported to prior releases.