Add baselines for unique_together for Django Evolution 0.7.
unique_together changed were ignored by Django Evolution in the past,
but that's no longer the case with 0.7. However, projects do need to
establish a baseline for unique_together now, to get the database back
in sync with the evolution signatures.

This adds a ChangeMeta for every unique_together we use, in order to
establish that baseline.