Update to CodeMirror 1.20, and enable auto-continuation of lists.
This updates CodeMirror to 1.20, which has some fixes for lists in
Markdown (in particular, bullet points at the beginning of a list
sometimes turning into emphasis).

It also enables auto-continuation of lists. For example, starting a line
with "* blah" and hitting enter will start the next line off with a new
list item. This works for ordered and unordered lists.