Use the new text_type API, and remove client-side Markdown escaping.
This makes use of the new `?force-text-type` capabilities of the API to
force all requests to come in as Markdown. It also ensures all
appropriate text rendered in the templates are Markdown, meaning that
the JavaScript doesn't have to deal with handling escaping. That greatly
simplifies the Markdown-related logic, removes inconsistencies in
escaping logic, and prevents browser-related issues with Regular
Expressions.