Add better support for archiving repositories.
Archiving had issues before where an archived repository's path would
still conflict with a newly created repository. This confused a lot of
people using RBCommons, and required some manual work on our end to fix.

Now, we store a flag indicating that a repository is archived, and we
store the archived timestamp. The timestamp becomes part of the
unique_together for the path, preventing that from conflicting.

Reviewed at https://reviews.reviewboard.org/r/6043/