Improve the usefulness of the user info box.
The user info box has always been less useful than it should be. It was
scared of mouse pointers, meaning it would disappear as soon as you
moved your mouse onto it, making it impossible to select e-mail
addresses.

Now it's possible to move the mouse onto the infobox and click the
username (going to their profile page) or the e-mail address (which is a
valid mail link).

The box will fade out once the mouse moves off the box, after a small
delay. It also fades in much faster now.