Add multi-select actions and visual improvements to the dashboard.
This is the first big update we've had to the dashboard in some time.

First, it adds support for multiple selection of review requests, and
actions that can be invoked for those that are selected. The actions
right now are limited to closing review requests, but more can be added
down the road.

The action pane appears over the sidebar. It's in a darker color and is
semi-transparent (but only barely). It goes away when no review requests
are selected.

This also makes the dashboard take up the full height of the page
(excluding some padding and the content above it). The rows of entries
scroll now, instead of having to scroll the entire page. With the new
changes to datagrids in Djblets, the rows are able to scroll without
affecting the header, meaning you can always see what column represents
what cell.