Add support for the Unfuddle hosting service.
This supports Unfuddle as a hosting service. Unfuddle provides Git and
Subversion hosting.

Unfuddle's pretty straightforward to support. It makes use of HTTP Basic
Auth, and requires base commit IDs for posting.

There's also some improvements to RepositoryForm to provide more data to
the HostingService's authorize and check_repository functions.
The Unfuddle support uses this to do better validation of a repository
and an account, which we should make use of for other services.