Add useful query parameters to filter lists of repositories.
This adds a handful of query parameters for filtering lists of
repositories by various criteria. This allow filtering by name,
path OR mirror path, name OR path OR mirror path, tool name, hosting
service ID, repository username OR hosting service username.

Each of these supports a comma-separated list of values, which work as
an OR.

We'll be able to use this in RBTools to make repository lookups faster.