Fix links for file attachment URLs on CDNs in e-mails.
File attachment URLs on CDNs weren't showing up correctly in e-mails. We
were prepending the http(s)://domain to them, which was a problem when
that was already included in the URL. We now guarantee that
FileAttachment.get_absolute_url will return an actual absolute URL,
domain included, and it will only prepend the http(s)://domain if it's
not already there.