Fix the dashboard's multi-selection with reloading.
There were a couple bugs causing multi-selection to break when
reloading.

First, the timer pausing logic wasn't working, due to a typo in a
variable name. This also resulted in toggling a checkbox causing
multiple timers to start and stack.

Second, after a reload, the events for the actions in the sidebar were
no longer hooked up. This is fixed by a simple delegateEvents() call.