Size the autocomplete list appropriately and keep it in view.
The autocomplete results list was using a default size of the input,
which was fine with the old review request layout, but was rather
squashed with the new one. It also would scroll off the size of the page
pretty easily.

We now give it a default width for the review request autocompletes that
extends out a bit, giving enough room for the contents.

The results list is also now smartly positioned to stay on-screen. It
will find a suitable edge to dock to ensure it can be seen in full.