Fix z-index issues for the AJAX spinner and headbar menus.
The z-index for the AJAX spinner caused it to appear below the review
dialog, which looked a bit wrong. It's now set to be higher.

The user and support menus were also appearing above everything else,
due to an attempt at fixing a z-index issue in the admin UI with grid
filters. Turns out we had a rule for the filters to set the right
z-index, but it was being overridden. We now force it, and no longer
have the custom z-index.