Stop showing empty space in the review request box.
On a blank or short review request, the review request box would show
the text areas and then empty space. This was pretty ugly.

We now fill in the remaining space with the last textarea (Testing Done
by default, though this should adapt to custom extension-provided text
fields when those are supported). It will take up the full height,
cleaning up the look.

When the field reaches a point where it'd otherwise extend past those
initial bounds, it will just behave as it did before, autosizing to fit
the content. This ends up feeling pretty natural. As you type in,
say, Description, Testing Done gets pushed down in order to accomodate
Description's size (as it did before), but the bottom now stays
anchored, up until the point where the field hits its natural size and
starts increasing the size of the review request box.