Fix user lookup in ReviewRequest/ReviewRequestDraft API resources.
The original intention was to retrieve the first available user in
auth.get_backends(). However, if an exception happens in the try clause,
it goes to the except which passes and then the variable "user" is not
initialized. The function get_or_create_user either returns an actual
user, None, or throws exception.

Reviewed at https://reviews.reviewboard.org/r/5189/