Fix building of static media for extensions.
The static media building support for extensions depended on a
settings_local.py file somewhere in the path, which we had due to
running setup.py develop on our dev environments, but users won't on any
site installs.

We now package a nearly blank settings_local.py file and use that for
the static media building.