Add support for publicly-accessible LocalSites.
LocalSites have always been locked down to members of the groups, but
there are advantages to allowing for anonymous access when simply used
for partitioning of a server by teams/organizations.

This adds a public flag to LocalSite. If set, non-members will be able
to access anything that's considered public within the LocalSite. This
essentially gives it the same behavior as a standard site.