Clean up the look of review UI pages.
Review UI pages were pretty ugly. They had this bit of the review
request box sticking out with nothing but the caption. This was from
back when we filled that with the screenshot list.

Now, by default, a review UI using the default.html template or the
`review_ui_content_box` block will have its UI shown in the confines of
that box. It has the whole area to play with, and isn't restricted by
any padding or UI elements. It will automatically have the actions
drop-down right above it as well.

Review UIs not using this template or block will appear below as before.
In this case, though, the bit of box will be removed, showing just the
action bar, cleanly floating there.

One side-effect of this is that the caption is no longer shown
automatically. It's now up to the review UI to decide how to show this.