Fix Markdown rendering issues with the new field support.
The new field support didn't convert new review requests to Markdown,
yet it would escape the text. The result was seeing an uploaded
description in escaped Markdown.

We want to always render the built-in Description and Testing Done
fields in Markdown. To do this, there's now an `always_render_markdown`
attribute on the field that ensures the result will always be Markdown.
This will still factor in whether the source text needs to be escaped or
not.