Fix issues with Git defaults and SVN file access for Beanstalk.
There were a couple Beanstalk-related issues that were hit in the wild.

The first is that the default repository path for Git repositories was
wrong. People generally didn't notice due to the REPOSITORY= entry in
.reviewboardrc, and since we don't actually use it for anything but
path-based matching.

The second, and more important, is that Subversion flat-out didn't work.
It was using the Beanstalk blob API, which throws an HTTP 500 for
Subversion. I thought this had worked in the past, but it's possible it
regressed. Either way, simple change to make the blob API Git-only.

While at it, I also improved the speed of file checks by not passing any
?contents= to the URL unless we did want contents. It seems the API is
currently treating ?contents=0 the same as ?contents=1.