Cache comment issue counts on ReviewRequest.
ReviewRequest now has three new counters for tracking the number of
issue states across all comments made on the review request.

These default to None for all existing review requests. When a review
request is accessed, all its counter fields will be initialized in one
go. Any time the state of an issue on a public review is changed,
the counter will be updated.

This will be used in future changes for the API and for the dashboard.