Fixed rendering changed field information in change descriptions.
The fields change over-escaped values when rendering. This was a
last-minute change made during review, due to the naming choices for the
functions and variables.

I've removed escaping for the change description values, which were all
rendering safely to HTML except for one. I've also updated the function
names to more explicitly state that they are providing valid HTML.