Add a security checklist in the admin site.
There are a few security issues that people have run into which are related to
the deployment of Review Board, and can't be fixed with a simple code change.
In most cases, we can make rb-site and our example configuration files do the
right thing for new deployments, but that doesn't help out people who are
upgrading.

Until now, we've solved this by writing up KB articles on how to fix the
configuration, and trying to advertise them heavily. This change introduces a
much nicer method: a page in the admin site which will automatically check for
misconfigurations which can affect server security, and suggest fixes when it
discovers one.

There are two checks currently implemented: executable state of uploaded files,
and the ALLOWED_HOSTS setting.

For uploaded files, if the web server is configured to automatically execute
certain file types (like .php), a user could upload such a file as an
attachment to a review request, and the web server would happily run the code
therein. This is fixable by changing the web server config to disable all
executable types within the uploaded media directory.

The ALLOWED_HOSTS django setting exists to explicitly list out the hostnames
for which django will respond. This defaults to '*', which is handy for
development, but potentially allows people to perform host poisoning attacks.

Testing Done:
Ran through all tests and they succeeded when they should and also failed when
they should when I introduced failure conditions.

Reviewed at https://reviews.reviewboard.org/r/4952/