Fix an overflow issue on the sidebar on the user page.
My recent fix to the user page resulted in the sidebar overflowing the
list of review requests when the list was shorter than the sidebar
contents.

This changes the rule for the sidebar on the user page to use a float
instead of absolute positioning. Absolute positioning is what we want
for the dashboard, due to the fact that we control the size of the
dashboard itself, but as that's not (currently?) the case on the user
page, we need let it use its natural height.