Provide raw downloading of files in the diff viewer.
The diff viewer now shows a download icon next to the revision
information for each file, allowing the original or new versions of a
file to be downloaded. This also works with files on interdiffs.

This can be used for easily downloading a patched file at any stage of a
diff revision range, or generating a diff based on the original and
modified files outside of Review Board.

Binary files use this as well instead of the previous "Download" links.
The existing actions lists remain, but won't be shown unless there's
content on either side. That can happen if there's a "Review" or
"New Comment" link.