Fix 'enter' in autocomplete fields (reviewboard)
Handling of the 'enter' key for autocomplete fields like people and groups
wasn't correct, because the djblets handler (which submits the field) was
running before the autocomplete handler (which populates the field with the
current selected item).

The fix is to ensure that the keydown handlers run in the other order. This
change does the review board side of that, using the new 'deferEventSetup'
option to avoid connecting the events until after we set up autocomplete.

Testing done:
Reproduced the bug. After this fix (and the matching reviewboard one), I can
press "enter" with a selected user and it will properly populate the field.

Fixes bug 2779.

Reviewed at https://reviews.reviewboard.org/r/5283/