Show open issues instead of Ship It! in the dashboard and reviews.
The dashboard now shows the number of open issues instead of the number
of Ship Its if there are open issues. If there are no open issues left,
the Ship Its are displayed instead. This makes it a lot easier to see at
a glance on the dashboard if a change is ready to go in, or if work
still needs to be done.

On a review box, "Ship it!" will be replaced by "Fix it, then Ship it!"
when there are open issues. As soon as they're all resolved for that
review, the text will be replaced by "Ship it!" Likewise, re-opening an
issue will flip the text back to "Fix it, then Ship It!"