Add the server info and capabilities to the root resource.
Nearly every client request RBTools makes needs to fetch both the root
resource and the server info resource.

The server info is pretty useful for most clients, especially the
capabilities and version information, so putting it in a separate
resource basically just results in one extra request.

Since the data is cheap and useful, it's now located in the root
resource. The server info resource still exists and mirrors the same
data, but clients can now check the root resource first.